Move functions in util.interpolation to util.coordinates
