--- climada/util/interpolation.py
@@ -1,325 +1,0 @@
-"""
-This file is part of CLIMADA.
-
-Copyright (C) 2017 ETH Zurich, CLIMADA contributors listed in AUTHORS.
-
-CLIMADA is free software: you can redistribute it and/or modify it under the
-terms of the GNU General Public License as published by the Free
-Software Foundation, version 3.
-
-CLIMADA is distributed in the hope that it will be useful, but WITHOUT ANY
-WARRANTY; without even the implied warranty of MERCHANTABILITY or FITNESS FOR A
-PARTICULAR PURPOSE.  See the GNU General Public License for more details.
-
-You should have received a copy of the GNU General Public License along
-with CLIMADA. If not, see <https://www.gnu.org/licenses/>.
-
----
-
-Define interpolation functions using different metrics.
-"""
-
-__all__ = ['interpol_index',
-           'dist_sqr_approx',
-           'DIST_DEF',
-           'METHOD']
-
-import logging
-import numpy as np
-import numba
-import scipy as sp
-
-from sklearn.neighbors import BallTree
-from climada.util.constants import ONE_LAT_KM, EARTH_RADIUS_KM
-
-LOGGER = logging.getLogger(__name__)
-
-DIST_DEF = ['approx', 'haversine', 'euclidean']
-"""Distances"""
-
-METHOD = ['NN']
-"""Interpolation methods"""
-
-THRESHOLD = 100
-"""Distance threshold in km. Nearest neighbors with greater distances are
-not considered."""
-
-@numba.njit
-def dist_approx(lats1, lons1, cos_lats1, lats2, lons2):
-    """Compute equirectangular approximation distance in km."""
-    d_lon = lons1 - lons2
-    d_lat = lats1 - lats2
-    return np.sqrt(d_lon * d_lon * cos_lats1 * cos_lats1 + d_lat * d_lat) * ONE_LAT_KM
-
-@numba.njit
-def dist_sqr_approx(lats1, lons1, cos_lats1, lats2, lons2):
-    """Compute squared equirectangular approximation distance. Values need
-    to be sqrt and multiplicated by ONE_LAT_KM to obtain distance in km."""
-    d_lon = lons1 - lons2
-    d_lat = lats1 - lats2
-    return d_lon * d_lon * cos_lats1 * cos_lats1 + d_lat * d_lat
-
-def interpol_index(centroids, coordinates, method=METHOD[0],
-                   distance=DIST_DEF[1], threshold=THRESHOLD,
-                   **kwargs):
-    """Returns for each coordinate the centroids indexes used for
-    interpolation.
-
-    Parameters
-    ----------
-    centroids : 2d array
-        First column contains latitude, second
-        column contains longitude. Each row is a geographic point
-    coordinates : 2d array
-        First column contains latitude, second
-        column contains longitude. Each row is a geographic point
-    method : str, optional
-        interpolation method to use. 'NN' default.
-    distance : str
-        distance to use. Options are defined in interpolation.DIST_DEF.
-        The default is 'euclidian'.
-    threshold : float, optional
-        distance threshold in km over which no neighbor will
-        be found. Those are assigned with a -1 index.
-        The default is 100 (km)
-    kwargs: dict, optional
-        keyword arguments to be passed on to the interpolation method
-
-    Returns
-    -------
-    numpy array with so many rows as coordinates containing the
-    centroids indexes
-
-    Notes
-    -----
-    The default method uses the 'euclidean' distance metric to
-    find the nearest neighbors ('NN'). This method is fast for (quasi-)gridded
-    data, but introduces innacuracy since distances in lat/lon coordinates
-    are not equal to distances in meters on the Earth surface, in particular
-    for higher latitude and distances larger than 100km. If more accuracy
-    is needed, please use 'haversine' distance metric. This however is slower
-    for (quasi-)gridded data.
-
-    """
-    if (method == METHOD[0]) & (distance == DIST_DEF[0]):
-        # Compute for each coordinate the closest centroid
-        interp = index_nn_aprox(centroids, coordinates, threshold, **kwargs)
-    elif (method == METHOD[0]) & (distance == DIST_DEF[1]):
-        # Compute the nearest centroid for each coordinate using the
-        # haversine formula. This is done with a Ball tree.
-        interp = index_nn_haversine(centroids, coordinates, threshold, **kwargs)
-    elif (method == METHOD[0]) & (distance == DIST_DEF[2]):
-        # Compute the nearest centroid for each coordinate using the
-        # euclidean distance on a kTree. This is best for gridded data.
-        interp = index_nn_euclidean(centroids, coordinates, threshold, **kwargs)
-    else:
-        raise ValueError(
-            f'Interpolation using {method} with distance {distance} is not supported.')
-    return interp
-
-def index_nn_aprox(centroids, coordinates, threshold=THRESHOLD,
-                   check_antimeridian=True):
-    """Compute the nearest centroid for each coordinate using the
-    euclidean distance d = ((dlon)cos(lat))^2+(dlat)^2. For distant points
-    (e.g. more than 100km apart) use the haversine distance.
-
-    Parameters
-    ----------
-    centroids : 2d array
-        First column contains latitude, second
-        column contains longitude. Each row is a geographic point
-    coordinates : 2d array
-        First column contains latitude, second
-        column contains longitude. Each row is a geographic point
-    threshold : float
-        distance threshold in km over which no neighbor will
-        be found. Those are assigned with a -1 index
-    check_antimedirian: bool
-        if True, recomputes the nn in a strip with lon size equal to threshold
-        around the antimeridian using the Haversine distance. The antimeridian
-        is guessed from both coordinates and centroids, and is assumed
-        equal to 0.5*(lon_max+lon_min) + 180.
-
-    Returns
-    -------
-    array with so many rows as coordinates containing the centroids indexes
-    """
-
-    # Compute only for the unique coordinates. Copy the results for the
-    # not unique coordinates
-    _, idx, inv = np.unique(coordinates, axis=0, return_index=True,
-                            return_inverse=True)
-    # Compute cos(lat) for all centroids
-    centr_cos_lat = np.cos(np.radians(centroids[:, 0]))
-    assigned = np.zeros(coordinates.shape[0], int)
-    num_warn = 0
-    for icoord, iidx in enumerate(idx):
-        dist = dist_sqr_approx(centroids[:, 0], centroids[:, 1],
-                               centr_cos_lat, coordinates[iidx, 0],
-                               coordinates[iidx, 1])
-        min_idx = dist.argmin()
-        # Raise a warning if the minimum distance is greater than the
-        # threshold and set an unvalid index -1
-        if np.sqrt(dist.min()) * ONE_LAT_KM > threshold:
-            num_warn += 1
-            min_idx = -1
-
-        # Assign found centroid index to all the same coordinates
-        assigned[inv == icoord] = min_idx
-
-    if num_warn:
-        LOGGER.warning('Distance to closest centroid is greater than %s'
-                       'km for %s coordinates.', threshold, num_warn)
-
-    if check_antimeridian:
-        assigned = index_antimeridian_strip(centroids,
-                                            coordinates,
-                                            threshold,
-                                            assigned)
-
-    return assigned
-
-def index_nn_haversine(centroids, coordinates, threshold=THRESHOLD):
-    """Compute the neareast centroid for each coordinate using a Ball
-    tree with haversine distance.
-
-    Parameters
-    ----------
-    centroids : 2d array
-        First column contains latitude, second
-        column contains longitude. Each row is a geographic point
-    coordinates : 2d array
-        First column contains latitude, second
-        column contains longitude. Each row is a geographic point
-    threshold : float
-        distance threshold in km over which no neighbor will
-        be found. Those are assigned with a -1 index
-
-
-    Returns
-    -------
-    array with so many rows as coordinates containing the centroids indexes
-    """
-    # Construct tree from centroids
-    tree = BallTree(np.radians(centroids), metric='haversine')
-    # Select unique exposures coordinates
-    _, idx, inv = np.unique(coordinates, axis=0, return_index=True,
-                            return_inverse=True)
-
-    # query the k closest points of the n_points using dual tree
-    dist, assigned = tree.query(np.radians(coordinates[idx]), k=1,
-                                return_distance=True, dualtree=True,
-                                breadth_first=False)
-
-    # Raise a warning if the minimum distance is greater than the
-    # threshold and set an unvalid index -1
-    num_warn = np.sum(dist * EARTH_RADIUS_KM > threshold)
-    if num_warn:
-        LOGGER.warning('Distance to closest centroid is greater than %s'
-                       'km for %s coordinates.', threshold, num_warn)
-        assigned[dist * EARTH_RADIUS_KM > threshold] = -1
-
-    # Copy result to all exposures and return value
-    return np.squeeze(assigned[inv])
-
-
-def index_nn_euclidean(centroids, coordinates, threshold=THRESHOLD,
-                       check_antimeridian=True):
-    """Compute the neareast centroid for each coordinate using a k-d tree.
-
-    Parameters
-    ----------
-    centroids : 2d array
-        First column contains latitude, second column contains longitude.
-        Each row is a geographic point
-    coordinates : 2d array
-        First column contains latitude, second column contains longitude. Each
-        row is a geographic point
-    threshold : float
-        distance threshold in km over which no neighbor will be found. Those
-        are assigned with a -1 index
-    check_antimedirian: bool
-        if True, recomputes the nn in a strip with lon size equal to threshold
-        around the antimeridian using the Haversine distance. The antimeridian
-        is guessed from both coordinates and centroids, and is assumed
-        equal to 0.5*(lon_max+lon_min) + 180.
-
-    Returns
-    -------
-    np.array
-        with so many rows as coordinates containing the centroids indexes
-    """
-    # Construct tree from centroids
-    tree = sp.spatial.KDTree(np.radians(centroids))
-    # Select unique exposures coordinates
-    _, idx, inv = np.unique(coordinates, axis=0, return_index=True,
-                            return_inverse=True)
-
-    # query the k closest points of the n_points using dual tree
-    dist, assigned = tree.query(np.radians(coordinates[idx]), k=1, p=2, workers=-1)
-
-    # Raise a warning if the minimum distance is greater than the
-    # threshold and set an unvalid index -1
-    num_warn = np.sum(dist * EARTH_RADIUS_KM > threshold)
-    if num_warn:
-        LOGGER.warning('Distance to closest centroid is greater than %s'
-                       'km for %s coordinates.', threshold, num_warn)
-        assigned[dist * EARTH_RADIUS_KM > threshold] = -1
-
-    if check_antimeridian:
-        assigned = index_antimeridian_strip(centroids,
-                                            coordinates[idx],
-                                            threshold,
-                                            assigned)
-
-    # Copy result to all exposures and return value
-    return np.squeeze(assigned[inv])
-
-def index_antimeridian_strip(centroids, coordinates, threshold, assigned):
-    """Recompute nearest neighbors close to the anti-meridian with the
-    Haversine distance
-
-    Parameters
-    ----------
-    centroids : 2d array
-        First column contains latitude, second column contains longitude.
-        Each row is a geographic point
-    coordinates : 2d array
-        First column contains latitude, second column contains longitude. Each
-        row is a geographic point
-    threshold : float
-        distance threshold in km over which no neighbor will be found. Those
-        are assigned with a -1 index
-    assigned : 1d array
-        coordinates that have assigned so far
-
-    Returns
-    -------
-    np.array
-        with so many rows as coordinates containing the centroids indexes
-
-    """
-    lon_min = min(centroids[:, 1].min(), coordinates[:, 1].min())
-    lon_max = max(centroids[:, 1].max(), coordinates[:, 1].max())
-    if lon_max - lon_min > 360:
-        raise ValueError("Longitudinal coordinates need to be normalized"
-                         "to a common 360 degree range")
-    mid_lon = 0.5 * (lon_max + lon_min)
-    antimeridian = mid_lon + 180
-
-    thres_deg = np.degrees(threshold / EARTH_RADIUS_KM)
-    coord_strip_bool = coordinates[:, 1] + antimeridian < 1.5 * thres_deg
-    coord_strip_bool |= coordinates[:, 1] - antimeridian >  -1.5 * thres_deg
-    if np.any(coord_strip_bool):
-        coord_strip = coordinates[coord_strip_bool]
-        cent_strip_bool = centroids[:, 1] + antimeridian < 2.5 * thres_deg
-        cent_strip_bool |= centroids[:, 1] - antimeridian >  -2.5 * thres_deg
-        if np.any(cent_strip_bool):
-            cent_strip = centroids[cent_strip_bool]
-            strip_assigned = index_nn_haversine(cent_strip,
-                                                coord_strip,
-                                                threshold)
-            new_coords = np.where(cent_strip_bool)[0][strip_assigned]
-            new_coords[np.where(strip_assigned == -1)] = -1
-            assigned[coord_strip_bool] = new_coords
-    return assigned
